fixed half stars not working

+++ src/server/restaurants-schema.sql
@@ -8,7 +8,7 @@ create table restaurants
     cuisine varchar(255),
     city varchar(255),
     state varchar(3),
-    rating integer,
+    rating decimal(2,1),
     image varchar(255),
     descrip varchar(1000));
 
@@ -17,5 +17,5 @@ create table reviews
     res_id integer references restaurants(id),
     rev_name varchar(30),
     rev_date varchar(255),
-    rating integer,
+    rating decimal(2,1),
     review varchar(1000));
